AIRLINES/requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
coverage==7.4.1
crashtest==0.4.1
cryptography==42.0.1
distlib==0.3.8
Django==5.0.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-stubs==3.14.5
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.2.0
msgpack==1.0.7
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
paramiko==3.4.0
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.4.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pyannotate==1.2.0
pycparser==2.21
PyMySQL==1.1.0
PyNaCl==1.5.0
pyproject_hooks==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.7.0
pytz==2023.3.post1
rapidfuzz==3.6.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
sshtunnel==0.1.4
tomlkit==0.12.3
trove-classifiers==2024.1.8
types-pytz==2023.4.0.20240130
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240125
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
xattr==0.10.1
zipp==3.17.0
